fix: resolve build failures and runtime errors across full stack
- Add missing ElasticCmd sources to manager Makefile (linker error)
- Enable Drogon PostgreSQL support via Conan options (runtime FATAL)
- Replace env var placeholders in config.prod.json with Docker values
  (Drogon doesn't support ${} interpolation, caused UInt parse error)
- Cast HttpStatusCode to int for spdlog/fmt compatibility
- Use std::string for SQL raw literals (ambiguous operator<< overload)
- Fix CardActionArea href prop (not in MUI API, use component='a')
- Fix s3-frontend Docker build context path (package.json not found)

+++ backend/src/services/ElasticClient.cpp
@@ -49,7 +49,8 @@ void ElasticClient::indexDoc(
                 onErr(502, "ES connection failed");
                 return;
             }
-            auto code = resp->getStatusCode();
+            auto code =
+                static_cast<int>(resp->getStatusCode());
             if (code >= 200 && code < 300) {
                 onOk(json::parse(resp->getBody(),
                                  nullptr, false));
@@ -80,7 +81,8 @@ void ElasticClient::search(
                 onErr(502, "ES connection failed");
                 return;
             }
-            auto code = resp->getStatusCode();
+            auto code =
+                static_cast<int>(resp->getStatusCode());
             if (code >= 200 && code < 300) {
                 onOk(json::parse(resp->getBody(),
                                  nullptr, false));
